homework_4/requirements.txt
flask
python-dotenv
grequests